13
Removed most of the OID work from main() and moved it to oidRead().
I still need to get main() to not set the class ifIndex and ifDescr to
the last one found.
I also need to get oidRead() to not set the class ifSpeed, ifInOctets
and ifOutOctets to the last one found.